+++ packages/mix_generator/lib/src/core/builders/spec_styler_class_builder.dart
@@ -1026,8 +1026,8 @@ class SpecStylerClassBuilder {
     return buildMixableSpecTargetCall(
       annotation: annotation,
       specElement: specElement,
-      specName: specName,
       stylerName: stylerName,
+      allowExactGeneratedStyler: true,
       hostElement: specElement,
       hostLibrary: specElement.library,
     );